+++ src/mpeg2_macroblock_blocks.rs
@@ -430,6 +430,7 @@ mod tests {
             macroblock_pattern: false,
             macroblock_intra: true,
             spatial_temporal_weight_code_flag: false,
+            spatial_temporal_weight_class: Some(0),
             bit_position_after: 0,
         }
     }
@@ -442,6 +443,7 @@ mod tests {
             macroblock_pattern: true,
             macroblock_intra: false,
             spatial_temporal_weight_code_flag: false,
+            spatial_temporal_weight_class: Some(0),
             bit_position_after: 0,
         }
     }
@@ -674,6 +676,7 @@ mod tests {
             macroblock_pattern: false,
             macroblock_intra: false,
             spatial_temporal_weight_code_flag: false,
+            spatial_temporal_weight_class: Some(0),
             bit_position_after: 0,
         };
         let cbp = CodedBlockPattern {